Restructure the AST
